+++ packages/api/src/controllers.rs
@@ -1,2 +1,6 @@
 pub mod health_check;
 pub mod not_found;
+
+pub mod totp {
+    pub mod create_totp;
+}
