+++ packages/api/src/controllers/totp/create_totp.rs
@@ -1,0 +1,22 @@
+use crate::utils::{connect_to_mongodb::MongoDB, get_env::get_env};
+use axum::{http::StatusCode, Extension, Json};
+use serde::Serialize;
+use std::sync::Arc;
+
+#[derive(Serialize)]
+pub struct CreateTotpResponse {
+    message: &'static str,
+    database: bool,
+    deployment_environment: String,
+}
+
+pub async fn create_totp(
+    mongodb: Extension<Arc<MongoDB>>,
+) -> (StatusCode, Json<CreateTotpResponse>) {
+    let response: CreateTotpResponse = CreateTotpResponse {
+        message: "TOTP",
+        database: mongodb.collection.find_one(None, None).await.is_ok(),
+        deployment_environment: get_env().NEXT_PUBLIC_DEPLOYMENT_ENVIRONMENT,
+    };
+    (StatusCode::OK, Json(response))
+}
