--- packages/api/src/handlers/totp.rs
@@ -1,1 +1,0 @@
-pub mod create_totp;
