+++ packages/api/src/services/totp.rs
@@ -1,0 +1,7 @@
+pub struct Totp;
+
+impl Totp {
+    pub async fn create_totp() -> String {
+        "Saul Goodman".to_string()
+    }
+}
